Add task stats section
