+++ app/controllers/raif/admin/application_controller.rb
@@ -15,6 +15,20 @@ module Raif
         end
       end
 
+      def get_time_range(period)
+        case period
+        when "day"
+          24.hours.ago..Time.current
+        when "week"
+          1.week.ago..Time.current
+        when "month"
+          1.month.ago..Time.current
+        when "all"
+          Time.at(0)..Time.current
+        else
+          24.hours.ago..Time.current
+        end
+      end
     end
   end
 end
