+++ app/controllers/raif/admin/stats_controller.rb
@@ -14,23 +14,6 @@ module Raif
         @conversation_entry_count = Raif::ConversationEntry.where(created_at: @time_range).count
         @agent_count = Raif::Agent.where(created_at: @time_range).count
       end
-
-    private
-
-      def get_time_range(period)
-        case period
-        when "day"
-          24.hours.ago..Time.current
-        when "week"
-          1.week.ago..Time.current
-        when "month"
-          1.month.ago..Time.current
-        when "all"
-          Time.at(0)..Time.current
-        else
-          24.hours.ago..Time.current
-        end
-      end
     end
   end
 end
