+++ app/views/raif/admin/stats/index.html.erb
@@ -54,7 +54,12 @@
                   <h6 class="mb-0 text-muted">Tasks</h6>
                 </div>
               </div>
-              <h3 class="fw-bold mb-0"><%= number_with_delimiter(@task_count) %></h3>
+              <div class="d-flex justify-content-between align-items-center">
+                <h3 class="fw-bold mb-0"><%= number_with_delimiter(@task_count) %></h3>
+                <div class="mt-2">
+                  <%= link_to t("raif.admin.common.details"), raif.admin_stats_tasks_path(period: @selected_period) %>
+                </div>
+              </div>
             </div>
           </div>
